added if then else to alias script

+++ u22/them.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 
-# this works from either the Home folder or the USB top level
-\cp -v HJS/u22/customer.bash_aliases ~/.bash_aliases
-cd ~/
+# this works from Home or USB, top level or inside repo folder
+if [ -f ./customer.bash_aliases ]; then
+    \cp -v ./customer.bash_aliases ~/.bash_aliases
+    printf "\n\e[1;35m   Restart terminal to load new aliases.\e[0m\n\n"
+elif [ -f HJS/u22/customer.bash_aliases ]; then
+    \cp -v HJS/u22/customer.bash_aliases ~/.bash_aliases
+    printf "\n\e[1;35m   Restart terminal to load new aliases.\e[0m\n\n"
+else
+    printf "\n\e[1;35m   Alias file not found.\e[0m\n\n"
+fi
 
-printf "\n\e[1;35m   Restart terminal to load new aliases.\e[0m\n\n"
+cd ~/
 
 # EOF
